Add function for write and read ADC registers. Made small changes in spi.c

+++ soft1/adc_receiver/src/drivers/adc16dv160.h
@@ -8,8 +8,8 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-void adc16dv160_write(uint8_t addr, uint16_t data);
-uint16_t adc16dv160_read(uint8_t addr);
+void adc16dv160_write(uint8_t addr, uint8_t data);
+uint8_t adc16dv160_read(uint8_t addr);
 
 #endif	/* _ADC16DV160_H_ */
